Improve annotation tag compiler-plugin validation
