+++ compiler-plugin/src/main/java/io/ballerina/stdlib/constraint/compiler/annotation/tag/AnnotationTagInt.java
@@ -18,32 +18,15 @@
 
 package io.ballerina.stdlib.constraint.compiler.annotation.tag;
 
-import java.util.ArrayList;
-
-import static io.ballerina.stdlib.constraint.compiler.Constants.CONSTRAINT_MAX_VALUE;
-import static io.ballerina.stdlib.constraint.compiler.Constants.CONSTRAINT_MAX_VALUE_EXCLUSIVE;
-import static io.ballerina.stdlib.constraint.compiler.Constants.CONSTRAINT_MIN_VALUE;
-import static io.ballerina.stdlib.constraint.compiler.Constants.CONSTRAINT_MIN_VALUE_EXCLUSIVE;
 import static io.ballerina.stdlib.constraint.compiler.Constants.TYPE_INT;
 
 /**
  * The class to represent the `@constraint:Int` annotation tag.
  */
-public class AnnotationTagInt implements AnnotationTag {
+public class AnnotationTagInt extends AbstractValueConstraints {
 
     @Override
     public boolean isCompatibleFieldType(String fieldType) {
         return fieldType.equals(TYPE_INT);
-    }
-
-    @Override
-    public boolean haveCompatibleConstraints(ArrayList<String> constraints) {
-        return !(constraints.contains(CONSTRAINT_MIN_VALUE) && constraints.contains(CONSTRAINT_MIN_VALUE_EXCLUSIVE) ||
-                constraints.contains(CONSTRAINT_MAX_VALUE) && constraints.contains(CONSTRAINT_MAX_VALUE_EXCLUSIVE));
-    }
-
-    @Override
-    public boolean isValidConstraintValue(double constraintValue) {
-        return true;
     }
 }
